Split native Zenmap topology view builders
